+++ src/component/OnlineCount.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Online({}) {
+export default function Online({} = {}) {
   return (
     <div style={{margin: 'auto', display: 'flex', alignItems: 'strect'}}>
       {/* <div>4,999 online users</div> */}