+++ src/component/TaskBar.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-export default function TaskBar({ start, length, fill }) {
+export default function TaskBar({ start, end, fill }) {
   return (
-    <rect x={start} y='0' width={length} height='2320' fill={fill} />
+    <rect x={start} y='0' width={end - start} height='2320' fill={fill} />
   )
 }